// Client/src/common/protobuf/game/room_pb.ts
// @generated by protoc-gen-es v2.2.0 with parameter "target=ts"
// @generated from file common/protobuf/game/room.proto (package room, syntax proto3)
/* eslint-disable */

import type { GenFile, GenMessage } from "@bufbuild/protobuf/codegenv1";
import { fileDesc, messageDesc } from "@bufbuild/protobuf/codegenv1";
import type { C2S_Metadata } from "../client/client_pb";
import { file_common_protobuf_client_client } from "../client/client_pb";
import type { S2C_Metadata } from "../server/server_pb";
import { file_common_protobuf_server_server } from "../server/server_pb";
import type { UserInfo } from "../common/struct_pb";
import { file_common_protobuf_common_struct } from "../common/struct_pb";
import type { Message } from "@bufbuild/protobuf";

/**
 * Describes the file common/protobuf/game/room.proto.
 */
export const file_common_protobuf_game_room: GenFile = /*@__PURE__*/
  fileDesc("Ch9jb21tb24vcHJvdG9idWYvZ2FtZS9yb29tLnByb3RvEgRyb29tIlMKDUMyTF9FbnRlclJvb20SIgoEbWV0YRgBIAEoCzIULmNsaWVudC5DMlNfTWV0YWRhdGESDgoGcm9vbUlkGAIgASgEEg4KBnVzZXJJZBgDIAEoCSJ3Cg1MMkNfRW50ZXJSb29tEiIKBG1ldGEYASABKAsyFC5zZXJ2ZXIuUzJDX01ldGFkYXRhEhEKCWlzRW50ZXJlZBgCIAEoCBIOCgZyb29tSWQYAyABKAQSHwoFdXNlcnMYBCADKAsyEC5jb21tb24uVXNlckluZm9iBnByb3RvMw", [file_common_protobuf_client_client, file_common_protobuf_server_server, file_common_protobuf_common_struct]);

/**
 * 방 입장 요청 패킷
 *
 * @generated from message room.C2L_EnterRoom
 */
export type C2L_EnterRoom = Message<"room.C2L_EnterRoom"> & {
  /**
   * @generated from field: client.C2S_Metadata meta = 1;
   */
  meta?: C2S_Metadata;

  /**
   * 입장하려는 Room ID
   *
   * @generated from field: uint64 roomId = 2;
   */
  roomId: bigint;

  /**
   * 유저 ID
   *
   * @generated from field: string userId = 3;
   */
  userId: string;
};

/**
 * Describes the message room.C2L_EnterRoom.
 * Use `create(C2L_EnterRoomSchema)` to create a new message.
 */
export const C2L_EnterRoomSchema: GenMessage<C2L_EnterRoom> = /*@__PURE__*/
  messageDesc(file_common_protobuf_game_room, 0);

/**
 * 방 입장 응답 패킷
 *
 * @generated from message room.L2C_EnterRoom
 */
export type L2C_EnterRoom = Message<"room.L2C_EnterRoom"> & {
  /**
   * 공통 응답 메타 데이터
   *
   * @generated from field: server.S2C_Metadata meta = 1;
   */
  meta?: S2C_Metadata;

  /**
   * 방 입장 성공 여부
   *
   * @generated from field: bool isEntered = 2;
   */
  isEntered: boolean;

  /**
   * 방 ID
   *
   * @generated from field: uint64 roomId = 3;
   */
  roomId: bigint;

  /**
   * 방에 있는 기존 유저 정보
   *
   * @generated from field: repeated common.UserInfo users = 4;
   */
  users: UserInfo[];
};

/**
 * Describes the message room.L2C_EnterRoom.
 * Use `create(L2C_EnterRoomSchema)` to create a new message.
 */
export const L2C_EnterRoomSchema: GenMessage<L2C_EnterRoom> = /*@__PURE__*/
  messageDesc(file_common_protobuf_game_room, 1);